# swagger.yaml
openapi: 3.0.3
info:
  title: Banas API
  description: |
    A comprehensive business management API for handling customers, daily entries, billing, payments, and route management.
    This API provides endpoints for managing a milk delivery business with features including:
    - Customer management and account tracking
    - Daily entry recording for deliveries
    - Automated billing system
    - Payment processing and due tracking
    - Route management for delivery optimization
    - QR code generation for customer verification
    - Dashboard analytics and reporting
  version: 1.0.0
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://banas.up.railway.app
    description: Production server
  - url: http://localhost:8000
    description: Development server
security:
  - BearerAuth: []
paths:
...
  /api/customer/:
    get:
      tags:
        - Customers
      summary: List all customers
      description: Retrieve all customers with pagination support
      parameters:
        - name: page
          in: query
          description: Page number
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          description: Number of items per page
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Customers retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  count:
                    type: integer
                  next:
                    type: string
                    nullable: true
                  previous:
                    type: string
                    nullable: true
                  results:
                    type: array
                    items:
                      $ref: '#/components/schemas/Customer'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
    
    post:
      tags:
        - Customers
      summary: Create a new customer
      description: Create a new customer account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerCreate'
      responses:
        '201':
          description: Customer created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /api/customer/{id}/:
    get:
      tags:
        - Customers
      summary: Get customer details
      description: Retrieve specific customer information
      parameters:
        - name: id
          in: path
          required: true
          description: Customer UUID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Customer details retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
    
    put:
      tags:
        - Customers
      summary: Update customer
      description: Update existing customer information
      parameters:
        - name: id
          in: path
          required: true
          description: Customer UUID
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerCreate'
      responses:
        '200':
          description: Customer updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...
  /api/dailyentry/{id}/:
    get:
      tags:
        - Daily Entries
      summary: Get daily entry details
      description: Retrieve specific daily entry information
      parameters:
        - name: id
          in: path
          required: true
          description: Daily Entry UUID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Daily entry retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DailyEntry'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
    
    delete:
      tags:
        - Daily Entries
      summary: Delete daily entry
      description: Remove a daily entry record
      parameters:
        - name: id
          in: path
          required: true
          description: Daily Entry UUID
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Daily entry deleted successfully
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /api/dailyentry/historical/:
    get:
      tags:
        - Daily Entries
      summary: Get historical data
      description: Retrieve historical daily entry data with date range filtering
      parameters:
        - name: from_date
          in: query
          required: true
          description: Start date (YYYY-MM-DD)
          schema:
            type: string
            format: date
        - name: to_date
          in: query
          required: true
          description: End date (YYYY-MM-DD)
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Historical data retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DailyEntry'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...
  /api/dailyentry/customer/dailyentry/{customer_id}:
    post:
      tags:
        - Daily Entries
      summary: Create customer QR daily entry
      description: Create daily entry using customer QR code
      parameters:
        - name: customer_id
          in: path
          required: true
          description: Customer UUID
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                cooler:
                  type: integer
                  minimum: 1
                  example: 2
                qr_pin:
                  type: integer
                  example: 1234
      responses:
        '201':
          description: QR daily entry created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DailyEntry'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'

  # Bill Management Endpoints
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login endpoint
  schemas:
...
    Customer:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        sequence_no:
          type: integer
          nullable: true
          example: 1
        first_name:
          type: string
          maxLength: 200
          example: "John"
        last_name:
          type: string
          maxLength: 200
          example: "Smith"
        phone_no:
          type: string
          pattern: '^[789]\d{9}$'
          nullable: true
          example: "9876543210"
        route:
          type: string
          format: uuid
          nullable: true
          example: "550e8400-e29b-41d4-a716-446655440001"
        email:
          type: string
          format: email
          nullable: true
          example: "john.smith@example.com"
        rate:
          type: integer
          example: 25
        date_added:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00Z"
        date_updated:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00Z"
        addedby:
          type: string
          maxLength: 100
          nullable: true
          example: "admin"
        updatedby:
          type: string
          maxLength: 100
          nullable: true
          example: "admin"
        active:
          type: boolean
          default: true
          example: true
    CustomerCreate:
      type: object
      required:
        - first_name
        - last_name
        - rate
      properties:
        sequence_no:
          type: integer
          nullable: true
        first_name:
          type: string
          maxLength: 200
          example: "John"
        last_name:
          type: string
          maxLength: 200
          example: "Smith"
        phone_no:
          type: string
          pattern: '^[789]\d{9}$'
          nullable: true
          example: "9876543210"
        route:
          type: string
          format: uuid
          nullable: true
          example: "550e8400-e29b-41d4-a716-446655440001"
        email:
          type: string
          format: email
          nullable: true
          example: "john.smith@example.com"
        rate:
          type: integer
          example: 25
        addedby:
          type: string
          maxLength: 100
          nullable: true
          example: "admin"
        active:
          type: boolean
          default: true
...
    DailyEntry:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        customer:
          type: string
          format: uuid
          nullable: true
          example: "550e8400-e29b-41d4-a716-446655440001"
        cooler:
          type: integer
          example: 2
        date_added:
          type: string
          format: date-time
          nullable: true
          example: "2024-01-15T10:30:00Z"
        addedby:
          type: string
          maxLength: 100
          nullable: true
          example: "admin"
        updatedby:
          type: string
          maxLength: 100
          nullable: true
          example: "admin"
...
    ErrorResponse:
      type: object
      properties:
        detail:
          type: string
          example: "Error message"
        code:
          type: string
          example: "error_code"
    ValidationError:
      type: object
      properties:
        field_name:
          type: array
          items:
            type: string
          example: ["This field is required."]
  responses:
    BadRequestError:
      description: Bad Request
      content:
        application/json:
          schema:
            oneOf:
              - $ref: '#/components/schemas/ErrorResponse'
              - $ref: '#/components/schemas/ValidationError'
    UnauthorizedError:
      description: Unauthorized - Invalid or missing authentication token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            detail: "Authentication credentials were not provided."
...
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            detail: "Not found."
...